backend/requirements.txt
Django>=2.0,<3.0
psycopg2>=2.7,<3.0
djangorestframework
djangorestframework-jwt
django-rest-auth>=0.9.3
django-allauth>=0.37.1
django-cors-headers
